dockerimages/docker/requirements_fsl.txt
pydicom==2.3.1
matplotlib==3.7.1
Pillow==10.2.0
webdriver-manager==4.0.0
numpy==1.22.3
pandas==1.3.5
openpyxl==3.1.2
scipy>=1.11.4,<1.15
reportlab==4.1.0
PyPDF2==2.12.1
rich
seaborn
